stop caching compact index spec results
We definitely can’t cache this regardless of the inputs; potentially we can cache the results into a hash based on gem_names, but let’s wait and profile it to see if we actually need to do that. None of the other indexes cache here, so it seems like it’s probably safe.